# backend/docs/openapi.yaml
openapi: "3.0.3"
info:
  title: StudyApp API
  version: "1.0.0"
  description: Scan-to-Quiz educational platform API
servers:
  - url: http://localhost:8080
    description: Development
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum:
            - UNAUTHORIZED
            - FORBIDDEN
            - NOT_FOUND
            - VALIDATION_ERROR
            - CONFLICT
            - RATE_LIMIT_EXCEEDED
            - FEATURE_NOT_AVAILABLE
            - SCAN_LIMIT_EXCEEDED
            - INTERNAL_ERROR
            - PAYMENT_FAILED
        message:
          type: string
        details:
          nullable: true
        trace_id:
          type: string
    MessageResponse:
      type: object
      properties:
        message:
          type: string
    StatusResponse:
      type: object
      properties:
        status:
          type: string
    UserInfo:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        email:
          type: string
        role:
          type: string
...
    RegisterRequest:
      type: object
      required: [name, email, password, password_confirm, accept_terms]
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        password_confirm:
          type: string
        phone:
          type: string
        accept_terms:
          type: boolean
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    RefreshTokenRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        expires_in:
          type: integer
          format: int64
        token_type:
          type: string
        user:
          $ref: "#/components/schemas/UserInfo"
...
    ForgotPasswordRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
    ResetPasswordRequest:
      type: object
      required: [token, new_password, new_password_confirm]
      properties:
        token:
          type: string
        new_password:
          type: string
        new_password_confirm:
          type: string
...
    UsersPage:
      type: object
      properties:
        items:
          type: array
          items:
            type: object
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    JobsPage:
      type: object
      properties:
        items:
          type: array
          items:
            type: object
        total:
          type: integer
    AICostSummary:
      type: object
      properties:
        total_cost_estimate:
          type: number
        total_requests:
          type: integer
        rows:
          type: array
          items:
            type: object
    AuditLogPage:
      type: object
      properties:
        items:
          type: array
          items:
            type: object
    ContentFlagsPage:
      type: object
      properties:
        items:
          type: array
          items:
            type: object
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    ValidationError:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
paths:
...
  /auth/register:
    post:
      summary: Register a new user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "409":
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      summary: Login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/refresh:
    post:
      summary: Refresh access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenRequest"
      responses:
        "200":
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/forgot-password:
    post:
      summary: Initiate password reset
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordRequest"
      responses:
        "200":
          description: Reset email sent if account exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /auth/reset-password:
    post:
      summary: Complete password reset
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordRequest"
      responses:
        "200":
          description: Password reset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /admin/users:
    get:
      summary: List users (admin)
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Users page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsersPage"
  /admin/users/{userID}/suspend:
    put:
      summary: Suspend a user (admin)
      parameters:
        - name: userID
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason:
                  type: string
      responses:
        "200":
          description: User suspended
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResponse"
  /admin/users/{userID}/unsuspend:
    put:
      summary: Unsuspend a user (admin)
      parameters:
        - name: userID
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User unsuspended
  /admin/jobs:
    get:
      summary: List scan jobs (admin)
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Jobs page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobsPage"
  /admin/jobs/{jobID}/retry:
    post:
      summary: Retry a failed scan job (admin)
      parameters:
        - name: jobID
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job queued for retry
        "409":
          description: Job not retryable
  /admin/ai-costs:
    get:
      summary: AI cost summary (admin)
      parameters:
        - name: from
          in: query
          schema:
            type: string
            format: date
        - name: to
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: AI costs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AICostSummary"
  /admin/audit-logs:
    get:
      summary: Audit log entries (admin)
      parameters:
        - name: userId
          in: query
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Audit logs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditLogPage"
  /admin/content-flags:
    get:
      summary: Content moderation flags (admin)
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Content flags
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentFlagsPage"
  /admin/content-flags/{flagID}/resolve:
    put:
      summary: Resolve a content flag (admin)
      parameters:
        - name: flagID
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [action]
              properties:
                action:
                  type: string
                  enum: [approved, removed]
                reason:
                  type: string
      responses:
        "200":
          description: Flag resolved